Refactor components to components folder
