+++ src/App.js
@@ -1,16 +1,16 @@
 import './App.css';
-import Header from './Header';
-import Home from './Home';
+import Header from './components/Header';
+import Home from './components/Home';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import Checkout from './Checkout';
-import Login from './Login';
+import Checkout from './components/Checkout';
+import Login from './components/Login';
 import { useEffect } from 'react';
 import { auth } from './firebase';
 import { useStateValue } from './StateProvider';
-import Payment from './Payment';
+import Payment from './components/Payment';
 import { loadStripe } from '@stripe/stripe-js';
 import { Elements } from '@stripe/react-stripe-js'; 
-import Orders from './Orders';
+import Orders from './components/Orders';
 
 // Run npm install only in Functions folder / Backend
