+++ src/components/Checkout.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import './Checkout.css';
 import Subtotal from './Subtotal';
-import { useStateValue } from './StateProvider';
+import { useStateValue } from '../StateProvider';
 import CheckoutProduct from './CheckoutProduct';
 
 function Checkout() {
