+++ src/components/Login.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { Link, useNavigate } from 'react-router-dom'
 import './Login.css'
 import { useState } from 'react';
-import { auth } from './firebase';
+import { auth } from '../firebase';
 import userEvent from '@testing-library/user-event';
 
 function Login() {
