+++ src/components/Orders.js
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react";
-import { db } from "./firebase";
+import { db } from "../firebase";
 import './Orders.css'
-import { useStateValue } from "./StateProvider";
+import { useStateValue } from "../StateProvider";
 import Order from "./Order";
 
 function Orders() {
